5-1_projects/01_slack_summary_to_line/requirements.txt
slack_sdk==3.27.1
python-dotenv==1.0.0
requests==2.31.0
openai==1.3.0
